requirements.txt
torch
transformers
flwr
flask
nibabel
numpy
Pillow
torchvision
h5py
watchdog
